+++ services/session/ipc/proxy/avsession_proxy.cpp
@@ -178,7 +178,7 @@ int32_t AVSessionProxy::GetPixelMapBuffer(AVMetaData& metaData, MessageParcel& d
     std::shared_ptr<AVSessionPixelMap> mediaPixelMap = metaData.GetMediaImage();
     if (mediaPixelMap != nullptr) {
         mediaImageBuffer = mediaPixelMap->GetInnerImgBuffer();
-        mediaImageLength = mediaImageBuffer.size();
+        mediaImageLength = static_cast<int>(mediaImageBuffer.size());
         metaData.SetMediaLength(mediaImageLength);
     }
 
@@ -187,7 +187,7 @@ int32_t AVSessionProxy::GetPixelMapBuffer(AVMetaData& metaData, MessageParcel& d
     std::shared_ptr<AVSessionPixelMap> avQueuePixelMap = metaData.GetAVQueueImage();
     if (avQueuePixelMap != nullptr) {
         avQueueImageBuffer = avQueuePixelMap->GetInnerImgBuffer();
-        avQueueImageLength = avQueueImageBuffer.size();
+        avQueueImageLength = static_cast<int>(avQueueImageBuffer.size());
         metaData.SetAVQueueLength(avQueueImageLength);
     }
